requirements.txt
opencv-python==4.11.0.86
mediapipe==0.10.18
pyyaml==6.0.1
numpy==1.26.4
serial==0.0.97